backend/requirements.txt
asgiref==3.2.10
boto3==1.14.41
boto3-stubs==1.14.41.0
botocore==1.17.41
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.1
docutils==0.15.2
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
Markdown==3.2.2
python-dateutil==2.8.1
pytz==2019.2
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
